server: further progress  - now parses the SocketConnectionResponse and makes the information accessible  - now the response of the addons runs in a separate thread  - updated response data-size to long  - added missing TODOs

+++ src/main/java/org/intellimate/izou/server/Response.java
@@ -15,6 +15,6 @@ public interface Response {
     int getStatus();
     Map<String, List<String>> getHeaders();
     String getContentType();
-    int getDataSize();
+    long getDataSize();
     InputStream getData();
 }
